Changing wrappers to analytics package name, adding predictive package
