+++ nzpyida/analytics/predictive/regression.py
@@ -11,8 +11,8 @@
 from typing import Dict
 from nzpyida.frame import IdaDataFrame
 from nzpyida.base import IdaDataBase
-from nzpyida.wrappers.utils import map_to_props, materialize_df, make_temp_table_name
-from nzpyida.wrappers.predictive_modeling import PredictiveModeling
+from nzpyida.analytics.utils import map_to_props, materialize_df, make_temp_table_name
+from nzpyida.analytics.predictive.predictive_modeling import PredictiveModeling
 
 
 class Regression(PredictiveModeling):
