+++ nzpyida/analytics/preparation.py
@@ -9,8 +9,8 @@
 # The full license is in the LICENSE file, distributed with this software.
 #-----------------------------------------------------------------------------
 from nzpyida.frame import IdaDataFrame
-from nzpyida.wrappers.utils import map_to_props, materialize_df, make_temp_table_name
-from nzpyida.wrappers.utils import get_auto_delete_context
+from nzpyida.analytics.utils import map_to_props, materialize_df, make_temp_table_name
+from nzpyida.analytics.utils import get_auto_delete_context
 
 
 def __prepare(proc: str, in_df: IdaDataFrame, params: dict, out_table: str=None) -> IdaDataFrame:
